Moved numbers_to_neutral to queries/num.py
