+++ queries/__init__.py
@@ -181,78 +181,6 @@ def parse_num(node: Node, num_str: str) -> float:
     return num or 0.0
 
 
-# Neutral gender form of numbers
-NUMBERS_NEUTRAL = {
-    "1": "eitt",
-    "2": "tvö",
-    "3": "þrjú",
-    "4": "fjögur",
-    "21": "tuttugu og eitt",
-    "22": "tuttugu og tvö",
-    "23": "tuttugu og þrjú",
-    "24": "tuttugu og fjögur",
-    "31": "þrjátíu og eitt",
-    "32": "þrjátíu og tvö",
-    "33": "þrjátíu og þrjú",
-    "34": "þrjátíu og fjögur",
-    "41": "fjörutíu og eitt",
-    "42": "fjörutíu og tvö",
-    "43": "fjörutíu og þrjú",
-    "44": "fjörutíu og fjögur",
-    "51": "fimmtíu og eitt",
-    "52": "fimmtíu og tvö",
-    "53": "fimmtíu og þrjú",
-    "54": "fimmtíu og fjögur",
-    "61": "sextíu og eitt",
-    "62": "sextíu og tvö",
-    "63": "sextíu og þrjú",
-    "64": "sextíu og fjögur",
-    "71": "sjötíu og eitt",
-    "72": "sjötíu og tvö",
-    "73": "sjötíu og þrjú",
-    "74": "sjötíu og fjögur",
-    "81": "áttatíu og eitt",
-    "82": "áttatíu og tvö",
-    "83": "áttatíu og þrjú",
-    "84": "áttatíu og fjögur",
-    "91": "níutíu og eitt",
-    "92": "níutíu og tvö",
-    "93": "níutíu og þrjú",
-    "94": "níutíu og fjögur",
-    "101": "hundrað og eitt",
-    "102": "hundrað og tvö",
-    "103": "hundrað og þrjú",
-    "104": "hundrað og fjögur",
-}
-
-HUNDREDS = ("tvö", "þrjú", "fjögur", "fimm", "sex", "sjö", "átta", "níu")
-
-
-def numbers_to_neutral(s: str) -> str:
-    """Convert integers within the string s to voice
-    representations using neutral gender, i.e.
-    4 -> 'fjögur', 21 -> 'tuttugu og eitt'"""
-
-    def convert(m):
-        match = m.group(0)
-        n = int(match)
-        prefix = ""
-        if 121 <= n <= 999 and 1 <= (n % 10) <= 4 and not 11 <= (n % 100) <= 14:
-            # A number such as 104, 223, 871 (but not 111 or 614)
-            if n // 100 == 1:
-                prefix = "hundrað "
-            else:
-                prefix = HUNDREDS[n // 100 - 2] + " hundruð "
-            n %= 100
-            if n <= 4:
-                # 'tvö hundruð og eitt', 'sjö hundruð og fjögur'
-                prefix += "og "
-            match = str(n)
-        return prefix + NUMBERS_NEUTRAL.get(match, match)
-
-    return re.sub(r"(\d+)", convert, s)
-
-
 def country_desc(cc: str) -> str:
     """Generate Icelandic description of being in a particular country
     with correct preposition and case e.g. 'á Spáni', 'í Þýskalandi'."""
